+++ app/src/main/java/com/example/practica4/ui/screen/Components.kt
@@ -1,0 +1,63 @@
+package com.example.practica4.ui.screen
+
+import androidx.compose.foundation.background
+import androidx.compose.foundation.clickable
+import androidx.compose.foundation.layout.Box
+import androidx.compose.foundation.layout.Row
+import androidx.compose.foundation.layout.Spacer
+import androidx.compose.foundation.layout.size
+import androidx.compose.foundation.layout.width
+import androidx.compose.foundation.shape.CircleShape
+import androidx.compose.material3.Text
+import androidx.compose.runtime.Composable
+import androidx.compose.ui.Alignment
+import androidx.compose.ui.Modifier
+import androidx.compose.ui.graphics.Color
+import androidx.compose.ui.text.font.FontWeight
+import androidx.compose.ui.unit.dp
+import androidx.compose.ui.unit.sp
+import com.example.practica4.viewmodel.BluetoothViewModel
+
+@Composable
+fun BluetoothConnectionStatusIndicator(connectionState: BluetoothViewModel.ConnectionState) {
+    Row(verticalAlignment = Alignment.CenterVertically) {
+        Text(text = "Estado:")
+        Spacer(modifier = Modifier.width(8.dp))
+        Box(
+            modifier = Modifier
+                .size(16.dp)
+                .background(
+                    color = when (connectionState) {
+                        is BluetoothViewModel.ConnectionState.Disconnected -> Color.Gray
+                        is BluetoothViewModel.ConnectionState.Connecting -> Color.Yellow
+                        is BluetoothViewModel.ConnectionState.Connected -> Color.Green
+                        is BluetoothViewModel.ConnectionState.Error -> Color.Red
+                    },
+                    shape = CircleShape
+                )
+        )
+    }
+}
+
+@Composable
+fun RoundButton(
+    text: String,
+    onClick: () -> Unit,
+    color: Color,
+    modifier: Modifier = Modifier
+) {
+    Box(
+        contentAlignment = Alignment.Center,
+        modifier = modifier
+            .size(100.dp)
+            .background(color, CircleShape)
+            .clickable(onClick = onClick)
+    ) {
+        Text(
+            text = text,
+            color = Color.White,
+            fontWeight = FontWeight.Bold,
+            fontSize = 16.sp
+        )
+    }
+}
